Commit
The .java files and the TreeModelData files for the Image Browsing &
Editing application designed for the Object-Oriented Development in
Java module coursework in my MSc Computing course.

This code was written and submitted long before I learned the
importance of commenting code properly.

Toolbar button icons are not included due to licensing.